Add sequence->collection to gen:sequence and make streams collections
This provides a way to assert that a sequence is also a collection,
which is usually true, and to fall back to a slow implementation in the
worst case. I'm conflicted about adding this, and if it becomes
frequently used, a better solution may need to be accounted for (maybe
just erroring when it isn't implemented instead of using a fallback
implementation with `apply`?).

closes #6